Added a numPointsToReturn point compensator for when we have multiple curves.
Added a numPointsToReturn point compensator for when we have multiple curves. What this means is that -- since we divide numPointsToReturn points evenly into the number of bezier curves we're attaching together, sometimes there is an uneven division -- we add a couple extra points when necessary, so that the number of points to return is proper.